fix user-mode warnings 4

+++ hyperdbg/hprdbgctrl/code/debugger/commands/debugging-commands/dt-struct.cpp
@@ -401,7 +401,7 @@ CommandDtShowDataBasedOnSymbolTypes(
                                          IsPhysicalAddress ? DEBUGGER_READ_PHYSICAL_ADDRESS : DEBUGGER_READ_VIRTUAL_ADDRESS,
                                          READ_FROM_KERNEL,
                                          TargetPid,
-                                         StructureSize,
+                                         (UINT32)StructureSize,
                                          &DtOptions);
     }
     else
