+++ hyperdbg/script-eval/code/ScriptEngineEval.c
@@ -1764,7 +1764,7 @@ ScriptEngineExecute(PGUEST_REGS                    GuestRegs,
         *Indx = *Indx + 1;
 
         SrcVal0    = GetValue(GuestRegs, ActionDetail, VariablesList, Src0, FALSE, StackBuffer, StackIndx, StackBaseIndx, StackTempBaseIndx);
-        *StackIndx = *StackIndx + SrcVal0;
+        *StackIndx = (UINT32)(*StackIndx + SrcVal0);
 
         break;
